Rewrote sync_html in pure Python (no more rsync!)
I can't guarantee that this works *exactly* like rsync. It is a lot more
complex and a lot more nuanced than I can possibly capture in 100 lines,
even with Python. However, I'm pretty sure it does what you would expect
in all but the edgiest of cases. For the intention it was being used for,
it should work just fine.